Finished refactoring file and folder structure
