+++ src/components/Child_C.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import "../styles/App.css";
 
 
-class Child_C extends React.Component<Child_P,object> {
+class Child_C extends React.Component<Child_P, object> {
     constructor(props: Child_P) {
         super(props);
         this.onChange_Response1 = this.onChange_Response1.bind(this);
@@ -21,15 +21,15 @@ class Child_C extends React.Component<Child_P,object> {
         return(
             <div className="">
                 <p className="Query_C">
-                {this.props.query}
+                    {this.props.query}
                 </p>
                 <input 
-                type="text" 
-                className={this.props.Response1_className}
-                autoFocus={true}
-                value={this.props.content}
-                onChange={this.onChange_Response1}
-                onKeyDown={this.onEvent_Response1}
+                    type="text" 
+                    className={this.props.Response1_className}
+                    autoFocus={true}
+                    value={this.props.content}
+                    onChange={this.onChange_Response1}
+                    onKeyDown={this.onEvent_Response1}
                 />
             </div>
         );
