+++ src/index.tsx
@@ -1,14 +1,14 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import App from './components/App';
-import registerServiceWorker from './registerServiceWorker';
-import './styles/index.css';
+import * as React from "react";
+import * as ReactDOM from "react-dom";
+import App from "./components/App";
+import registerServiceWorker from "./registerServiceWorker";
+import "./styles/index.css";
 
 
 
 ReactDOM.render(
-  <App  />,
-  document.getElementById('root') as HTMLElement
+  <App />,
+  document.getElementById("root") as HTMLElement
 );
